feat: viewフォルダをなくしてcomponentsフォルダをwidgetフォルダに変更
