+++ lib/auth/auth.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:location_social_media/auth/login_or_register.dart';
-import 'package:location_social_media/view/pages/home_page.dart';
+import 'package:location_social_media/pages/home_page.dart';
 
 class AuthPage extends StatelessWidget {
   const AuthPage({super.key});
