+++ lib/pages/basics_example.dart
@@ -1,58 +1,49 @@
-// Copyright 2019 Aleksander Woźniak
 // SPDX-License-Identifier: Apache-2.0
 
 import 'package:flutter/material.dart';
 import 'package:table_calendar/table_calendar.dart';
 
-import '../../Utils/utils.dart';
+import '../Utils/utils.dart';
 
-class TableRangeExample extends StatefulWidget {
+class TableBasicsExample extends StatefulWidget {
   @override
-  _TableRangeExampleState createState() => _TableRangeExampleState();
+  _TableBasicsExampleState createState() => _TableBasicsExampleState();
 }
 
-class _TableRangeExampleState extends State<TableRangeExample> {
+class _TableBasicsExampleState extends State<TableBasicsExample> {
   CalendarFormat _calendarFormat = CalendarFormat.month;
-  RangeSelectionMode _rangeSelectionMode = RangeSelectionMode.toggledOn;
   DateTime _focusedDay = DateTime.now();
   DateTime? _selectedDay;
-  DateTime? _rangeStart;
-  DateTime? _rangeEnd;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('TableCalendar - Range'),
+        title: Text(""),
+        backgroundColor: Colors.transparent,
       ),
       body: TableCalendar(
+        //日本語対応
+        locale: 'Ja-JP',
+        //今の年分と月を中央に寄せる
+        headerStyle:
+            HeaderStyle(formatButtonVisible: false, titleCentered: true),
         firstDay: kFirstDay,
         lastDay: kLastDay,
+        // フォーカスしている日付
         focusedDay: _focusedDay,
-        selectedDayPredicate: (day) => isSameDay(_selectedDay, day),
-        rangeStartDay: _rangeStart,
-        rangeEndDay: _rangeEnd,
+        // カレンダーのフォーマット
         calendarFormat: _calendarFormat,
-        rangeSelectionMode: _rangeSelectionMode,
+        selectedDayPredicate: (day) {
+          return isSameDay(_selectedDay, day);
+        },
         onDaySelected: (selectedDay, focusedDay) {
           if (!isSameDay(_selectedDay, selectedDay)) {
             setState(() {
               _selectedDay = selectedDay;
               _focusedDay = focusedDay;
-              _rangeStart = null; // Important to clean those
-              _rangeEnd = null;
-              _rangeSelectionMode = RangeSelectionMode.toggledOff;
             });
           }
-        },
-        onRangeSelected: (start, end, focusedDay) {
-          setState(() {
-            _selectedDay = null;
-            _focusedDay = focusedDay;
-            _rangeStart = start;
-            _rangeEnd = end;
-            _rangeSelectionMode = RangeSelectionMode.toggledOn;
-          });
         },
         onFormatChanged: (format) {
           if (_calendarFormat != format) {
