+++ lib/pages/login_page.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:location_social_media/view/components/custom_button.dart';
-import 'package:location_social_media/view/components/custom_text_field.dart';
+import 'package:location_social_media/widget/custom_button.dart';
+import 'package:location_social_media/widget/custom_text_field.dart';
 
 class LoginPage extends StatefulWidget {
   final Function()? onTap;
