+++ lib/pages/splash_page.dart
