+++ lib/router/app_routes.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:location_social_media/view/pages/splash_page.dart';
+import 'package:location_social_media/pages/splash_page.dart';
 
 final rootNavigatorKey = GlobalKey<NavigatorState>();
