+++ lib/widget/comment.dart
@@ -1,8 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:location_social_media/view/components/delete_button.dart';
-import 'package:location_social_media/view/components/like_button.dart';
+import 'package:location_social_media/widget/delete_button.dart';
+import 'package:location_social_media/widget/like_button.dart';
 
 class Comment extends StatefulWidget {
   final String text;
